+++ Smith-suggestions.txt
@@ -1,10 +1,9 @@
-AdminBase contians the ability to put a user in local storage
+Done  -- AdminBase contians the ability to put a user in local storage
 so that can go with the other login stuff in login.js then delete
 adminbase.js for efficency and just organizatianlly from the dev standpoint
 
 
-I believe Admin.js in the login folder can be made into a dialouge because it is a button function and render like 
-delete and upload so could be moved and renamed
+I believe Admin.js in the login folder can be made into a login because it is page render for the login
 
 For all dialogs Tristan wants to move away from material-ui so we need to work in a new tool there 
 for design
